(installer) Change the used project for netifaces.
At start we were using the netifaces project. The main issue now is that
it does not provide any whl file and we need to installe GigaOctets of
developing tools to build the module.

We tried to build our own build whl but we failed.

The netifaces2 project ( https://github.com/SamuelYvon/netifaces-2 ) is
maintained and it provides whl.

The good part is that is that the API are even.